Move db folder inside lib
