+++ src/lib/vocab-actions.ts
@@ -1,7 +1,7 @@
 'use server';
 
-import { db } from '@/db';
-import { VocabItem, vocabulary } from '@/db/schema';
+import { db } from '@/lib/db';
+import { VocabItem, vocabulary } from '@/lib/db/schema';
 import { getLogger } from '@/utils/logger';
 import { eq } from 'drizzle-orm/sqlite-core/expressions';
 import { revalidatePath } from 'next/cache';
